累计多次更新： 实现了Modbus传JSON 对modbus EVENT进行互斥量包裹 提取出所有callback 将文件名修改为模块

+++ Master/UCOS3_STM32H7_compiler6.10/MIDDLELAYER/method.cpp
@@ -34,7 +34,7 @@ void Method::ClearStringBuff(char* str, int length)
 {
     if (str == NULL)
         return;
-	for(u8 i = 0; i < length; i++)
+	for(int i = 0; i < length; i++)
 	{
 		str[i] = '\0';
 	}
